+++ website/infra/editor/config/fileUploadUtil.js
@@ -118,8 +118,16 @@ export const getVideoToUpload = (url, thumbnailUrl) => {
   return videoWithRelativeUrl;
 };
 
-export const mockCustomAudioUploadFunc = (updateEntity, removeEntity) => {
+export const mockCustomAudioUploadFunc = (
+  updateEntity,
+  removeEntity,
+  fromSettings,
+  componentData
+) => {
   console.log('consumer wants to upload custom audio');
+  if (fromSettings) {
+    return mockImageUploadFunc(undefined, false, updateEntity, undefined, componentData);
+  }
   const audioToUpload = getAudioToUpload('f0f74f_35a1cdce4973490eac49e74c3244364d');
   setTimeout(() => {
     updateEntity({ data: audioToUpload });